requirements.txt
parsedatetime
plotly
dataframe_image
hopsworks
hsml
joblib
matplotlib
pandas
scikit-learn
seaborn
python-dotenv
requests
alpha_vantage
great_expectations
prophet
tensorflow
pandas_market_calendars
streamlit
streamlit_folium
asyncio
nest_asyncio
opencv-python-headless
easyocr
Pillow
python-dotenv
